# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@echoes-chat.com
    name: API Support
  description: This is a chat application API with WebSocket support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Echoes Chat API
  version: "1.0"
paths:
...
  /api/v1/ws/chat:
    get:
      parameters:
      - description: JWT Token
        in: query
        name: token
        required: true
        type: string
      responses: {}
      summary: WebSocket endpoint for real-time chat
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"